// src/api/analysis/types.ts
export interface ProblemAnalysisResponseDto {
    timeComplexity: string;
    timeComplexityReasoning: string;
    spaceComplexity: string;
    spaceComplexityReasoning: string;
    algorithmType: string;
    algorithmTypeReasoning: string;
    dataStructures: string;
    dataStructuresReasoning: string;
    solutionImplementation: string;
    solutionImplementationReasoning: string;
}